+++ config/chief.php
@@ -118,4 +118,13 @@ return [
      * 'quill' and 'redactor'. By default the free quill editor is set.
      */
     'editor' => 'quill',
+
+    /**
+     * Here you can define the stack of checks used by the healthmonitor.
+     *
+     */
+    'healthMonitor' => [
+        Thinktomorrow\Chief\HealthMonitor\Checks\HomepageSetCheck::class,
+        Thinktomorrow\Chief\HealthMonitor\Checks\HomepageAccessibleCheck::class,
+    ]
 ];